Convert era_commons to typescript
