+++ cypress/component/ERACommons/era_commons.spec.tsx
@@ -1,15 +1,22 @@
-import ERACommons from 'src/components/era_commons/ERACommons.jsx'
-import { decodeNihToken } from 'src/utils/ERACommonsUtils.js'
-import { AuthenticateNIH } from 'src/libs/ajax/AuthenticateNIH.js'
-import { User } from 'src/libs/ajax/User.js'
-import { Storage } from 'src/libs/storage.js'
+import ERACommons from 'src/components/era_commons/ERACommons'
+import { decodeNihToken } from 'src/utils/ERACommonsUtils'
+import { AuthenticateNIH } from 'src/libs/ajax/AuthenticateNIH'
+import { User } from 'src/libs/ajax/User'
+import { Storage } from 'src/libs/storage'
 import { mount } from 'cypress/react'
 import React from 'react'
 import { Buffer } from 'buffer'
 
-// Example token from https://broad-shibboleth-prod.appspot.com/dev/login?return-url=%2Fexample-return%3Ftoken%3D%3Ctoken%3E
-const encodedToken
-  = `eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJ
+interface Researcher {
+  id: number
+  firstName: string
+  lastName: string
+  email: string
+  eraCommonsId?: string
+  properties?: Array<{ propertyKey: string, propertyValue: string | number }>
+}
+
+const encodedToken = `eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJ
 lcmFDb21tb25zVXNlcm5hbWUiOiJ0ZXN0aW5nIiw
 iaWF0IjoxNzA5NTYxMjc2LCJleHAiOjE3MDk1NjQ
 4NzZ9.HKnqebvJVL63jKZDqheYWRSNaLvB92b0Dk
@@ -22,14 +29,14 @@ KNCf8PHBSwiotFxBCcDM-eAakIobz-XjJ7uAYM_Z
 M-C5Sq4vsz3jwhUWBmf8_J55bFAhM_n-AooxphpO
 bEc9G_sDJjolIqSjt_UwgT1JFPFA`
 
-const researcher = {
+const researcher: Researcher = {
   id: 1,
   firstName: 'Test',
   lastName: 'User',
   email: 'test@email.com',
 }
 
-const buildToken = (username, iat, exp) => {
+const buildToken = (username: string, iat: number, exp: number): string => {
   const part1 = `{"alg":"RS256","typ":"JWT"}`
   const part2 = {
     eraCommonsUsername: username,
@@ -45,9 +52,9 @@ describe('ERA Commons Utility', function () {
     const token = { 'nih-username-token': encodedToken }
     const decoded = await decodeNihToken(token)
     expect(decoded).to.not.equal(null)
-    expect(decoded.eraCommonsUsername).to.equal('testing')
-    expect(decoded.iat).to.equal(1709561276)
-    expect(decoded.exp).to.equal(1709564876)
+    expect(decoded!.eraCommonsUsername).to.equal('testing')
+    expect(decoded!.iat).to.equal(1709561276)
+    expect(decoded!.exp).to.equal(1709564876)
   })
   it('decodeNihToken returns null if the token is invalid', async function () {
     const token = { 'nih-username-token': 'invalid' }
@@ -61,13 +68,12 @@ describe('ERA Commons Component', function () {
     cy.stub(User, 'getMe').returns(researcher)
     mount(
       <ERACommons
-        isAuthorized={false}
-        className=""
         destination=""
-        header={true} // Triggers the NIH eRA Commons header
+        header={true}
         onNihStatusUpdate={() => {}}
-        required={true} // Triggers the required flag on the NIH eRA Commons ID
-        style={{}}
+        required={true}
+        location={undefined}
+        validationError={false}
       />,
     )
     cy.get('#era-commons-id').should('exist')
@@ -78,27 +84,24 @@ describe('ERA Commons Component', function () {
   })
 
   it('renders a populated ERA Commons component after having authenticated with NIH', function () {
-    const clonedResearcher = JSON.parse(JSON.stringify(researcher))
+    const clonedResearcher: Researcher = { ...researcher }
     clonedResearcher.eraCommonsId = 'testing'
     const iat = new Date().getTime()
-    const exp = iat + (30 * 24 * 60 * 60 * 1000) // iat + 30 days
+    const exp = iat + (30 * 24 * 60 * 60 * 1000)
     clonedResearcher.properties = [
       { propertyKey: 'eraAuthorized', propertyValue: 'true' },
       { propertyKey: 'eraExpiration', propertyValue: exp },
     ]
     cy.stub(User, 'getMe').returns(clonedResearcher)
-    // The saveNihUsr function only returns the user's properties, not the entire user object
     cy.stub(AuthenticateNIH, 'saveNihUsr').returns(clonedResearcher.properties)
     mount(
       <ERACommons
-        isAuthorized={true}
-        className=""
         destination=""
         location={{ search: `?nih-username-token=${buildToken('testing', iat, exp)}` }}
         header={true}
         onNihStatusUpdate={() => {}}
         required={true}
-        style={{}}
+        validationError={false}
       />,
     )
     cy.get('#era-commons-id').should('exist')
@@ -112,45 +115,39 @@ describe('ERA Commons Component', function () {
   it('shows an error when auth token decoding fails', function () {
     mount(
       <ERACommons
-        isAuthorized={false}
-        className=""
         destination=""
-        header={true} // Triggers the NIH eRA Commons header
+        header={true}
         onNihStatusUpdate={() => {}}
-        required={true} // Triggers the required flag on the NIH eRA Commons ID
-        style={{}}
+        required={true}
+        validationError={false}
         location={{ search: '?nih-username-token=invalid' }}
       />,
     )
-
-    // cy.stub(AuthenticateNIH, 'getECMProviderAuthUrl').throws(new Error('error'));
     cy.get('[data-cy=era-commons-authenticate-link]').should('exist')
     cy.get('[data-cy=era-commons-authenticate-link]').click()
     cy.get('[data-cy=era-commons-error-span]').should('be.visible')
   })
 
   it('shows an error when removing linked account fails', function () {
-    const eraAuthedUser = Object.assign({}, researcher, {
+    const eraAuthedUser: Researcher = {
+      ...researcher,
       eraCommonsId: 'testing',
       properties: [
         { propertyKey: 'eraAuthorized', propertyValue: 'true' },
-        { propertyKey: 'eraExpiration', propertyValue: Date.now() + (30 * 24 * 60 * 60 * 1000) }, // iat + 30 days
+        { propertyKey: 'eraExpiration', propertyValue: Date.now() + (30 * 24 * 60 * 60 * 1000) },
       ],
-    })
+    }
     cy.stub(User, 'getMe').returns(eraAuthedUser)
     mount(
       <ERACommons
-        isAuthorized={true}
-        className=""
         destination=""
-        header={true} // Triggers the NIH eRA Commons header
+        header={true}
         onNihStatusUpdate={() => {}}
-        required={true} // Triggers the required flag on the NIH eRA Commons ID
-        style={{}}
+        required={true}
+        validationError={false}
         location={{ search: '?nih-username-token=' + encodedToken }}
       />,
     )
-
     cy.stub(AuthenticateNIH, 'deleteAccountLinkage').throws(new Error('error'))
     cy.get('[data-cy=era-delete-icon]').should('exist')
     cy.get('[data-cy=era-delete-icon]').click({ force: true })
@@ -158,21 +155,17 @@ describe('ERA Commons Component', function () {
   })
 
   it('shows an error when ECM fails', function () {
-    // Enable RAS
     cy.stub(Storage, 'getEnv').returns('dev')
     cy.stub(Storage, 'getCurrentUser').returns(researcher)
     mount(
       <ERACommons
-        isAuthorized={false}
-        className=""
         destination=""
-        header={true} // Triggers the NIH eRA Commons header
+        header={true}
         onNihStatusUpdate={() => {}}
-        required={true} // Triggers the required flag on the NIH eRA Commons ID
-        style={{}}
+        required={true}
+        validationError={false}
       />,
     )
-
     cy.stub(AuthenticateNIH, 'getECMProviderAuthUrl').throws(new Error('error'))
     cy.get('[data-cy=era-commons-authenticate-link]').should('exist')
     cy.get('[data-cy=era-commons-authenticate-link]').click()
